Added Service functions to read rdf files and insert into database

+++ migrations/20200518152140-create-book-meta.js
@@ -6,7 +6,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER
+        type: Sequelize.INTEGER(6)
       },
       language: {
         type: Sequelize.STRING
